requirements.txt
asgiref==3.5.0
black==22.1.0
boto3==1.21.8
botocore==1.24.8
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==4.0.2
django-crispy-forms==1.14.0
django-filter==21.1
django-heroku==0.3.1
django-rest-registration==0.7.0
django-storages==1.12.3
djangorestframework==3.13.1
gunicorn==20.1.0
hashids==1.3.1
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jmespath==0.10.0
MarkupSafe==2.1.0
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
Pillow==9.0.1
platformdirs==2.5.1
psycopg2==2.9.3
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.8
whitenoise==6.0.0
